+++ delivery-service/src/main/java/com/winner/client/deliveryservice/deliverymanagerhub/presentation/dto/requests/DeliveryManagerHubRegistrationRequest.java
@@ -1,4 +1,4 @@
-package com.winner.client.deliveryservice.deliverymanagerhub.presentation.requests;
+package com.winner.client.deliveryservice.deliverymanagerhub.presentation.dto.requests;
 
 import com.winner.client.deliveryservice.deliverymanagerhub.application.dto.commnad.DeliveryManagerHubRegistrationCommand;
 import java.util.UUID;
